some javadoc
