+++ test/test-launcher/src/main/java/es/weso/shexsjava/Validate.java
@@ -20,7 +20,7 @@ import es.weso.utils.eitherios.EitherIOUtils;
 import scala.Option;
 
 /**
-
+ *
  * @author Jose Emilio Labra Gayo
  *
  */
